refactor CodyTool and related types for better organization
This commit refactors the CodyTool and related types to improve code organization and reduce circular dependencies.

- Moves `CodyToolConfig` and `ToolStatusCallback` interfaces to a new `types.ts` file.
- Updates imports to reflect the new location of these types.
- Creates an `ICodyTool` interface to define the Cody tool interface.
- Implements the `ICodyTool` interface in the `CodyTool` abstract class.
- Moves the `McpToolImpl` class to `CodyTool.ts` and implements the `ICodyTool` interface.
- Updates the `CodyToolFactory` to use the new `CodyToolConfig` type.
- Mocks dependencies in `CodyTool.test.ts` to isolate tests.
