+++ vscode/src/chat/agentic/CodyToolProvider.test.ts
@@ -5,9 +5,10 @@ import { URI } from 'vscode-uri'
 import * as openctxAPI from '../../../../lib/shared/src/context/openctx/api'
 import { mockLocalStorage } from '../../services/LocalStorageProvider'
 import type { ContextRetriever } from '../chat-view/ContextRetriever'
-import { CodyTool, type CodyToolConfig } from './CodyTool'
+import { CodyTool } from './CodyTool'
 import { CodyToolProvider, TestToolFactory, type ToolConfiguration } from './CodyToolProvider'
 import { toolboxManager } from './ToolboxManager'
+import { CodyToolConfig } from './types'
 
 const localStorageData: { [key: string]: unknown } = {}
 mockLocalStorage({
